Move file_outline to assistant_tool crate
