+++ crates/assistant_tool/src/assistant_tool.rs
@@ -1,4 +1,5 @@
 mod action_log;
+pub mod outline;
 mod tool_registry;
 mod tool_schema;
 mod tool_working_set;
